feat: add StaffDetailsPage for read-only staff member view with editable pay codes section
